scripts/database/requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.3.0
tabulate>=0.9.0
